+++ src/main/java/it/wsolutions/demotdd/service/AccountNotFoundError.java
@@ -2,7 +2,7 @@ package it.wsolutions.demotdd.service;
 
 public class AccountNotFoundError extends RuntimeException {
 
-  public AccountNotFoundError(int accountId) {
+  public AccountNotFoundError(long accountId) {
     super("No Account found with accountId : "+accountId);
   }
 }
