+++ src/main/java/it/wsolutions/demotdd/service/api/AccountService.java
@@ -4,5 +4,5 @@ import it.wsolutions.demotdd.model.BankAccount;
 
 public interface AccountService {
 
-  BankAccount getBankAccount(int accountId);
+  BankAccount getBankAccount(long accountId);
 }
